sharedvolume/requirements.txt
Flask
PyYAML  # ==3.13